+++ silhouette/app/com/mohiva/play/silhouette/impl/providers/oauth1/secrets/CookieSecret.scala
@@ -191,4 +191,4 @@ case class CookieSecretSettings(
   cookieDomain: Option[String] = None,
   secureCookie: Boolean = Play.isProd, // Default to sending only for HTTPS in production, but not for development and test.
   httpOnlyCookie: Boolean = true,
-  expirationTime: Duration = 5 minutes)
+  expirationTime: FiniteDuration = 5 minutes)